extract InputViewHolder from Input model
